cleanup/reorganize stats transport config code
