+++ src/main/java/org/jitsi/videobridge/stats/StatsManagerBundleActivator.java
@@ -55,66 +55,11 @@ public class StatsManagerBundleActivator
     public static final int DEFAULT_STAT_INTERVAL = 1000;
 
     /**
-     * The default value for statistics transport.
-     */
-    private static final String DEFAULT_STAT_TRANSPORT = null;
-
-    /**
-     * The <tt>Logger</tt> used by the <tt>StatsManagerBundleActivator</tt>
-     * class and its instances to print debug information.
-     */
-    private static final Logger logger
-        = Logger.getLogger(StatsManagerBundleActivator.class);
-
-    /**
-     * The name of the property which specifies the name of the PubSub node that
-     * will receive the statistics about the Videobridge if PubSub transport is
-     * used to send statistics.
-     */
-    public static final String PUBSUB_NODE_PNAME
-        = "org.jitsi.videobridge.PUBSUB_NODE";
-
-    /**
-     * The name of the property which specifies the name of the service that
-     * will receive the statistics about the Videobridge if PubSub transport is
-     * used to send statistics.
-     */
-    public static final String PUBSUB_SERVICE_PNAME
-        = "org.jitsi.videobridge.PUBSUB_SERVICE";
-
-    /**
-     * The value for callstats.io statistics transport.
-     */
-    public static final String STAT_TRANSPORT_CALLSTATS_IO = "callstats.io";
-
-    /**
-     * The value for COLIBRI statistics transport.
-     */
-    public static final String STAT_TRANSPORT_COLIBRI = "colibri";
-
-    /**
-     * The value for PubSub statistics transport.
-     */
-    public static final String STAT_TRANSPORT_PUBSUB = "pubsub";
-
-    /**
-     * The value used to enable the MUC statistics transport.
-     */
-    public static final String STAT_TRANSPORT_MUC = "muc";
-
-    /**
      * The name of the property which specifies the interval in milliseconds for
      * sending statistics about the Videobridge.
      */
     public static final String STATISTICS_INTERVAL_PNAME
         = "org.jitsi.videobridge.STATISTICS_INTERVAL";
-
-    /**
-     * The name of the property which specifies the transport for sending
-     * statistics about the Videobridge.
-     */
-    private static final String STATISTICS_TRANSPORT_PNAME
-        = "org.jitsi.videobridge.STATISTICS_TRANSPORT";
 
     protected ConfigProperty<Boolean> statsEnabled = StatsEnabledProperty.getInstance();
 
@@ -138,132 +83,6 @@ public class StatsManagerBundleActivator
      * started.
      */
     private ServiceRegistration<StatsManager> serviceRegistration;
-
-    /**
-     * Adds a new {@code StatsTransport} to a specific {@code StatsManager}. The
-     * newly-added {@code StatsTransport} in to repeatedly send
-     * {@code Statistics} at a specific {@code interval}.
-     *
-     * @param statsMgr the {@code StatsManager} to add the new
-     * {@code StatsTransport} to
-     * @param cfg the {@code ConfigurationService} to read property values from
-     * or {@code null} to read the property values from {@code System}
-     * @param interval the interval/period in milliseconds at which the
-     * newly-initialized and added {@code StatsTransport} is to repeatedly send
-     * {@code Statistics}
-     * @param transport the identifier of the {@code StatsTransport} to
-     * initialize and add to {@code statsMgr}
-     */
-    private void addTransport(
-            StatsManager statsMgr,
-            ConfigurationService cfg,
-            int interval,
-            String transport)
-    {
-        StatsTransport t = null;
-
-        if (STAT_TRANSPORT_CALLSTATS_IO.equalsIgnoreCase(transport))
-        {
-            t = new CallStatsIOTransport();
-        }
-        else if (STAT_TRANSPORT_COLIBRI.equalsIgnoreCase(transport))
-        {
-            t = new ColibriStatsTransport();
-        }
-        else if (STAT_TRANSPORT_PUBSUB.equalsIgnoreCase(transport))
-        {
-            Jid service;
-            try
-            {
-                service = JidCreate.from(cfg.getString(PUBSUB_SERVICE_PNAME));
-            }
-            catch (XmppStringprepException e)
-            {
-                logger.error("Invalid pubsub service name", e);
-                return;
-            }
-
-            String node = cfg.getString(PUBSUB_NODE_PNAME);
-            if(service != null && node != null)
-            {
-                t = new PubSubStatsTransport(service, node);
-            }
-            else
-            {
-                logger.error(
-                        "No configuration properties for PubSub service"
-                            + " and/or node found.");
-            }
-        }
-        else if (STAT_TRANSPORT_MUC.equalsIgnoreCase(transport))
-        {
-            logger.info("Using a MUC stats transport");
-            t = new MucStatsTransport();
-        }
-        else
-        {
-            logger.error(
-                    "Unknown/unsupported statistics transport: " + transport);
-        }
-
-        if (t != null)
-        {
-            // Each StatsTransport type/identifier (i.e. specified by the
-            // transport method argument) is allowed its own interval/period.
-            interval
-                = ConfigUtils.getInt(
-                        cfg,
-                        STATISTICS_INTERVAL_PNAME + "." + transport,
-                        interval);
-
-            // The interval/period of the Statistics better be the same as the
-            // interval/period of the StatsTransport.
-            if (statsMgr.findStatistics(VideobridgeStatistics.class, interval)
-                    == null)
-            {
-                statsMgr.addStatistics(new VideobridgeStatistics(), interval);
-            }
-
-            statsMgr.addTransport(t, interval);
-        }
-    }
-
-    /**
-     * Populates a specific {@code StatsManager} with newly-initialized
-     * {@code StatTransport}s as selected through {@code ConfigurationService}
-     * and/or {@code System} properties.
-     *
-     * @param statsMgr the {@code StatsManager} to populate with new
-     * {@code StatsTransport}s
-     * @param cfg the {@code ConfigurationService} to read property values from
-     * or {@code null} to read the property values from {@code System}
-     * @param interval the interval/period in milliseconds at which the
-     * newly-initialized and added {@code StatsTransport}s to repeatedly send
-     * {@code Statistics}
-     */
-    private void addTransports(
-            StatsManager statsMgr,
-            ConfigurationService cfg,
-            int interval)
-    {
-        String transports
-            = ConfigUtils.getString(
-                    cfg,
-                    STATISTICS_TRANSPORT_PNAME,
-                    DEFAULT_STAT_TRANSPORT);
-
-        if (transports == null || transports.length() == 0)
-        {
-            // It is OK to have the statistics enabled without explicitly
-            // choosing transports because the statistics may be exposed through
-            // the REST API as well.
-            return;
-        }
-
-        // Allow multiple transports.
-        for (String transport : transports.split(","))
-            addTransport(statsMgr, cfg, interval, transport);
-    }
 
     /**
      * Starts the <tt>StatsManager</tt> OSGi bundle in a <tt>BundleContext</tt>.
@@ -339,6 +158,7 @@ public class StatsManagerBundleActivator
         statsMgr.addStatistics(new VideobridgeStatistics(), interval);
 
         //TODO: support individual intervals for each transport
+        // Add StatsTransports to StatsManager.
         statsTransports.get().forEach(statsTransport -> {
             statsMgr.addTransport(statsTransport, interval);
 
@@ -349,9 +169,6 @@ public class StatsManagerBundleActivator
                 statsMgr.addStatistics(new VideobridgeStatistics(), interval);
             }
         });
-
-        // Add StatsTransports to StatsManager.
-//        addTransports(statsMgr, cfg, interval);
 
         statsMgr.start(bundleContext);
